[SpringBootApplication] 함수로뺐던걸 클래스로 빼고보니 main()이 어디서 많이 본 것 같은 형태로 변경됨!
