+++ src/main/java/com/nahwasa/practice/tobyspringboot/TobySpringBootApplication.java
@@ -1,12 +1,10 @@
 package com.nahwasa.practice.tobyspringboot;
 
 import org.springframework.boot.web.embedded.tomcat.TomcatServletWebServerFactory;
-import org.springframework.boot.web.server.WebServer;
 import org.springframework.boot.web.servlet.server.ServletWebServerFactory;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.web.context.support.AnnotationConfigWebApplicationContext;
 import org.springframework.web.servlet.DispatcherServlet;
 
 @Configuration
@@ -23,27 +21,7 @@ public class TobySpringBootApplication {
     }
 
     public static void main(String[] args) {
-        run(TobySpringBootApplication.class, args);
-    }
-
-    private static void run(Class<?> applicationClass, String... args) {
-        AnnotationConfigWebApplicationContext applicationContext = new AnnotationConfigWebApplicationContext() {
-            @Override
-            protected void onRefresh() {
-                super.onRefresh();
-
-                ServletWebServerFactory servletWebServerFactory = this.getBean(ServletWebServerFactory.class);
-                DispatcherServlet dispatcherServlet = this.getBean(DispatcherServlet.class);
-
-                WebServer webServer = servletWebServerFactory.getWebServer(servletContext -> {
-                    servletContext.addServlet("dispatcherServlet", dispatcherServlet)
-                            .addMapping("/*");
-                });
-                webServer.start();
-            }
-        };
-        applicationContext.register(applicationClass);
-        applicationContext.refresh();
+        MySpringApplication.run(TobySpringBootApplication.class, args);
     }
 
 }
